Version 0.0.2: Move OAuth settings from widget form to admin settings page
Version 0.0.2
The OAuth settings need setting only once (unless the user changes
something fairly drastic), so they have been moved from the widget
back-end form to a new settings page in the WP admin settings menu.
The widget form will focus on only those properties that affect the
display of the widget.
The username option has been removed as it is not necessary.