Apparently fixes strange behaviour of widget on Appearances > Widgets
No code after the widget title was being displayed on the Appearance >
Widgets page.  Some tweaks to the code fixed it locally, and a few more
tweaks fixed it online.  I’m not really sure what the problem actually
is, but I’m hoping it’s going to keep on working.
Renamed a few functions to make them more blipper-widget-specific.
Also made a few cosmetic changes to the display of the widget.