Fixes loading issues
Remove test for is_active_widget before loading dependencies and
creating the settings object.  This function only returns true if the
widget is on a sidebar.  When this widget plugin is first activated, it
isn’t on a sidebar, so is_active_widget returns false, and the
dependencies and settings aren’t loaded.  The failure to load the
settings means that the settings page isn’t created and all manner of
things go wrong.
Minor: reworded some of the user messages/instructions; tidied up the
code a little.