Removes client id and client secret
The client id and client secret are not needed for the widget, so I
have removed them.
The function that creates the Polaroid|Blipfoto client has been
simplified: it did special stuff if the client had already been
created, but it turned out that the client always had to be created
from scratch anyway, so this was removed.
Minor: reformatting the comments.