Adds uninstallation code to tidy up on plug-in deletion
Added: uninstall.php to clean up the database after the plugin is
deleted.  Used the WP-Spamshield plug-in as a guide as to how to do it.
Also:
Replaced: screenshot-3.png with a screenshot using the default
twenty-fifteen theme with no modifications.
Fixed: weird truncated plug-in description text.
Removed: commented-out client id and client secret, which were left
over from the last commit.