Includes JS Underscore library; updates latest compatible WP version
Without the Javascript Underscore library being loaded, the WP Color
Picker in the widget form wouldn’t necessarily work.  Luckily, in these
cases, the WP Color Picker fails gracefully, using a text box instead.
The text box is now formatted in the same way as the other text boxes
in the widget form.
The version of WP that Blipper Widget is now compatible up to is 4.4.
Reformatted some of the function documentation to match those in the
Blipper Widget class.